changed codes on index.html file under categories
