+++ src/main/java/TechDetectives/HobbyFinder/models/Category.java
@@ -7,50 +7,33 @@ import java.util.List;
 import java.util.Objects;
 
 @Entity
-public class Category {
+public class Category extends AbstractEntity {
 
-    @Id
-    @GeneratedValue
-    private int id;
-
-    private String name;
+    public String location;
 
     @OneToMany
-    @JoinColumn(name= "category_id")
-    private final List<Hobby> hobbies= new ArrayList<>();
+    @JoinColumn(name = "category_id")
+    private final List<Hobby> hobbies = new ArrayList<>();
 
     public Category() {
 
     }
 
 
-    public Category(String name) {
-
-        this.name = name;
-
-    }
-    public int getId() {
-        return id;
+    public Category(String location) {
+        super();
+        this.location = location;
     }
 
-    public String getName() {
-        return name;
+    public String getLocation() {
+        return location;
     }
 
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Category category = (Category) o;
-        return id == category.id;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
+    public void setLocation(String location) {
+        this.location = location;
     }
 }
+
+
+
+
