+++ src/main/resources/templates/fragments.html
@@ -5,7 +5,7 @@
 
     <!-- Bootstrap stylesheets and script -->
     <link th:href="@{/css/bootstrap.css}" rel="stylesheet" />
-    <link th:href="@{/css/techjobs.css}" rel="stylesheet" />
+    <link th:href="@{/css/styles.css}" rel="stylesheet" />
     <script type="text/javascript" th:src="@{/js/bootstrap.js}"></script>
 
     <title th:text="'Hobbies' + ${title == null ? '' : ' :: ' + title}">Hobbies</title>
